requirements.txt
Django==1.10.2
Pillow==3.4.1
Pygments==2.1.3
argparse==1.2.1
astdump==4.3
astmonkey==0.1.1
astor==0.5
astunparse==1.4.0
autopep8==1.2.4
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.1
click==6.6
codegen==1.0
coverage==4.2
decorator==4.0.10
django-contrib-comments==1.7.3
django-mptt==0.8.6
django-pytest==0.2.0
django-tagging==0.4.5
django-xmlrpc==0.1.7
djangorestframework==3.5.0
enum34==1.1.6
ipython==5.1.0
ipython-genutils==0.1.0
more-itertools==2.2
mots-vides==2015.5.11
pathlib2==2.1.0
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
portalocker==1.0.0
prompt-toolkit==1.0.7
ptyprocess==0.5.1
py==1.4.31
pydot==1.2.3
pyparsing==2.1.10
pytest==3.0.3
pytest-django==3.0.0
pytz==2016.7
regex==2016.9.22
simplegeneric==0.8.1
six==1.10.0
-e git+https://github.com/alifanov/smother.git#egg=smother
traitlets==4.3.1
ujson==1.35
unidiff==0.5.2
wcwidth==0.1.7
wheel==0.29.0
psycopg2
uwsgi
